Range dispatch predicate sorting and comments/documentation

+++ core/src/main/java/org/geysermc/geyser/registry/mappings/predicate/ItemRangeDispatchProperty.java
@@ -50,14 +50,14 @@ public enum ItemRangeDispatchProperty implements PredicateReader<ItemPredicateCo
         this.reader = reader;
     }
 
-    ItemRangeDispatchProperty(PredicateCreator<ItemPredicateContext, Double> creator) {
-        this((element, context) -> creator.create(readThreshold(element, context)));
+    ItemRangeDispatchProperty(PredicateCreator<ItemPredicateContext, Integer> creator) {
+        this((element, context) -> creator.create((int) readThreshold(element, context)));
     }
 
-    ItemRangeDispatchProperty(PredicateCreator<ItemPredicateContext, Double> creator, PredicateCreator<ItemPredicateContext, Double> normalised) {
+    ItemRangeDispatchProperty(PredicateCreator<ItemPredicateContext, Integer> creator, PredicateCreator<ItemPredicateContext, Double> normalised) {
         this((element, context) -> {
             double threshold = readThreshold(element, context);
-            return normalised(element, context) ? normalised.create(threshold) : creator.create(threshold);
+            return normalised(element, context) ? normalised.create(threshold) : creator.create((int) threshold);
         });
     }
